Filter windows by the amount of "rare" k-mers

+++ src/seq/aln.rs
@@ -117,7 +117,7 @@ impl Alignment {
     /// Creates a new Alignment from the record.
     pub fn from_record(record: &Record, contigs: Rc<ContigNames>) -> Self {
         let cigar = Cigar::infer_ext_cigar(record, ());
-        let contig_id = ContigId::new(record.tid() as u16);
+        let contig_id = ContigId::new(record.tid());
         let start = u32::try_from(record.pos()).unwrap();
         let ref_interval = if cigar.is_empty() {
             assert!(record.is_unmapped(), "Read is mapped, but has empty CIGAR!");
